[refactor] 감상 기록 / bindingAdapter의 package 변경 (#304)
