+++ android/app/src/main/java/com/teamtripdraw/android/ui/common/bindingAdapter/CommonBindingAdapter.kt
@@ -3,7 +3,6 @@ package com.teamtripdraw.android.ui.common.bindingAdapter
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
-import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.databinding.BindingAdapter
 import com.bumptech.glide.Glide
 import com.teamtripdraw.android.R
@@ -22,14 +21,4 @@ fun TextView.setMessageVisibility(isVisible: Boolean) {
     if (isVisible) {
         this.visibility = View.VISIBLE
     }
-}
-
-@BindingAdapter("app:setPostWritingImageVisibility")
-fun ConstraintLayout.setPostWritingImageVisibility(isVisible: Boolean) {
-    visibility =
-        if (isVisible) {
-            View.VISIBLE
-        } else {
-            View.GONE
-        }
 }
